+++ obstore/python/obstore/__init__.py
@@ -1,7 +1,6 @@
 from typing import TYPE_CHECKING
 
 from .__get import GetOptions, OffsetRange, SuffixRange
-from .__list import ListChunkType, ListResult, ObjectMeta
 from .__put import PutMode, PutResult, UpdateVersion
 from .__sign import HTTP_METHOD
 from ._attributes import Attribute, Attributes
