+++ obstore/python/obstore/_get.pyi
@@ -1,9 +1,9 @@
 from collections.abc import Sequence
 
 from .__get import GetOptions
-from .__list import ObjectMeta
 from ._attributes import Attributes
 from ._bytes import Bytes
+from ._list import ObjectMeta
 from .store import ObjectStore
 
 class GetResult:
@@ -106,6 +106,7 @@ class BytesStream:
         file is large, it may exceed the default timeout of 30 seconds. In this case,
         you may see an error like:
 
+        x
         ```
         GenericError: Generic {
             store: "HTTP",
