+++ obstore/python/obstore/_obstore.pyi
@@ -21,7 +21,10 @@ from ._get import get_ranges as get_ranges
 from ._get import get_ranges_async as get_ranges_async
 from ._head import head as head
 from ._head import head_async as head_async
+from ._list import ListChunkType as ListChunkType
+from ._list import ListResult as ListResult
 from ._list import ListStream as ListStream
+from ._list import ObjectMeta as ObjectMeta
 from ._list import list as list  # noqa: A004
 from ._list import list_with_delimiter as list_with_delimiter
 from ._list import list_with_delimiter_async as list_with_delimiter_async
